+++ Sources/TuistKit/Services/BuildService.swift
@@ -10,6 +10,7 @@ import TuistSupport
 enum BuildServiceError: FatalError {
     case schemeNotFound(scheme: String, existing: [String])
     case schemeWithoutBuildableTargets(scheme: String)
+    case buildProductsNotFound(path: AbsolutePath)
 
     // Error description
     var description: String {
@@ -18,6 +19,8 @@ enum BuildServiceError: FatalError {
             return "Couldn't find scheme \(scheme). The available schemes are: \(existing.joined(separator: ", "))."
         case let .schemeWithoutBuildableTargets(scheme):
             return "The scheme \(scheme) cannot be built because it contains no buildable targets."
+        case let .buildProductsNotFound(path):
+            return "The expected build products at \(path.pathString) were not found."
         }
     }
 
@@ -28,6 +31,8 @@ enum BuildServiceError: FatalError {
             return .abort
         case .schemeWithoutBuildableTargets:
             return .abort
+        case .buildProductsNotFound:
+            return .bug
         }
     }
 }
@@ -37,17 +42,23 @@ final class BuildService {
     let generator: Generating
 
     /// Xcode build controller.
-    let xcodebuildController: XcodeBuildControlling
+    let xcodeBuildController: XcodeBuildControlling
+
+    /// Locator for finding `xcodebuild` output directory.
+    let xcodeProjectBuildDirectoryLocator: XcodeProjectBuildDirectoryLocating
 
     /// Build graph inspector.
     let buildGraphInspector: BuildGraphInspecting
 
-    init(generator: Generating = Generator(contentHasher: CacheContentHasher()),
-         xcodebuildController: XcodeBuildControlling = XcodeBuildController(),
-         buildGraphInspector: BuildGraphInspecting = BuildGraphInspector())
-    {
+    init(
+        generator: Generating = Generator(contentHasher: CacheContentHasher()),
+        xcodeBuildController: XcodeBuildControlling = XcodeBuildController(),
+        xcodeProjectBuildDirectoryLocator: XcodeProjectBuildDirectoryLocating = XcodeProjectBuildDirectoryLocator(),
+        buildGraphInspector: BuildGraphInspecting = BuildGraphInspector()
+    ) {
         self.generator = generator
-        self.xcodebuildController = xcodebuildController
+        self.xcodeBuildController = xcodeBuildController
+        self.xcodeProjectBuildDirectoryLocator = xcodeProjectBuildDirectoryLocator
         self.buildGraphInspector = buildGraphInspector
     }
 
@@ -56,6 +67,7 @@ final class BuildService {
         generate: Bool,
         clean: Bool,
         configuration: String?,
+        buildOutputPath: AbsolutePath?,
         path: AbsolutePath
     ) throws {
         let graph: ValueGraph
@@ -70,17 +82,34 @@ final class BuildService {
 
         logger.log(level: .debug, "Found the following buildable schemes: \(buildableSchemes.map(\.name).joined(separator: ", "))")
 
+        let workspacePath = try buildGraphInspector.workspacePath(directory: path)!
+
         if let schemeName = schemeName {
             guard let scheme = buildableSchemes.first(where: { $0.name == schemeName }) else {
                 throw BuildServiceError.schemeNotFound(scheme: schemeName, existing: buildableSchemes.map(\.name))
             }
-            try buildScheme(scheme: scheme, graphTraverser: graphTraverser, path: path, clean: clean, configuration: configuration)
+
+            try buildScheme(
+                scheme: scheme,
+                graphTraverser: graphTraverser,
+                workspacePath: workspacePath,
+                clean: clean,
+                configuration: configuration,
+                buildOutputPath: buildOutputPath
+            )
         } else {
             var cleaned: Bool = false
             // Run only buildable entry schemes when specific schemes has not been passed
             let buildableEntrySchemes = buildGraphInspector.buildableEntrySchemes(graphTraverser: graphTraverser)
             try buildableEntrySchemes.forEach {
-                try buildScheme(scheme: $0, graphTraverser: graphTraverser, path: path, clean: !cleaned && clean, configuration: configuration)
+                try buildScheme(
+                    scheme: $0,
+                    graphTraverser: graphTraverser,
+                    workspacePath: workspacePath,
+                    clean: !cleaned && clean,
+                    configuration: configuration,
+                    buildOutputPath: buildOutputPath
+                )
                 cleaned = true
             }
         }
@@ -90,21 +119,73 @@ final class BuildService {
 
     // MARK: - private
 
-    private func buildScheme(scheme: Scheme, graphTraverser: GraphTraversing, path: AbsolutePath, clean: Bool, configuration: String?) throws {
+    private func buildScheme(
+        scheme: Scheme,
+        graphTraverser: GraphTraversing,
+        workspacePath: AbsolutePath,
+        clean: Bool,
+        configuration: String?,
+        buildOutputPath: AbsolutePath?
+    ) throws {
         logger.log(level: .notice, "Building scheme \(scheme.name)", metadata: .section)
         guard let (project, target) = buildGraphInspector.buildableTarget(scheme: scheme, graphTraverser: graphTraverser) else {
             throw BuildServiceError.schemeWithoutBuildableTargets(scheme: scheme.name)
         }
-        let workspacePath = try buildGraphInspector.workspacePath(directory: path)!
+
         let buildArguments = buildGraphInspector.buildArguments(project: project, target: target, configuration: configuration, skipSigning: false)
-        _ = try xcodebuildController.build(
-            .workspace(workspacePath),
-            scheme: scheme.name,
-            clean: clean,
-            arguments: buildArguments
+
+        _ = try xcodeBuildController
+            .build(
+                .workspace(workspacePath),
+                scheme: scheme.name,
+                clean: clean,
+                arguments: buildArguments
+            )
+            .printFormattedOutput()
+            .toBlocking()
+            .last()
+
+        if let buildOutputPath = buildOutputPath {
+            let configuration = configuration ?? project.settings.defaultDebugBuildConfiguration()?.name ?? BuildConfiguration.debug.name
+            try copyBuildProducts(
+                to: buildOutputPath,
+                workspacePath: workspacePath,
+                platform: target.platform,
+                configuration: configuration
+            )
+        }
+    }
+
+    private func copyBuildProducts(
+        to outputPath: AbsolutePath,
+        workspacePath: AbsolutePath,
+        platform: Platform,
+        configuration: String
+    ) throws {
+        let xcodeSchemeBuildPath = try xcodeProjectBuildDirectoryLocator.locate(
+            platform: platform,
+            projectPath: workspacePath,
+            configuration: configuration
         )
-        .printFormattedOutput()
-        .toBlocking()
-        .last()
+        guard FileHandler.shared.exists(xcodeSchemeBuildPath) else {
+            throw BuildServiceError.buildProductsNotFound(path: xcodeSchemeBuildPath)
+        }
+
+        let buildOutputPath = outputPath.appending(component: xcodeSchemeBuildPath.basename)
+        if !FileHandler.shared.exists(buildOutputPath) {
+            try FileHandler.shared.createFolder(buildOutputPath)
+        }
+        logger.log(level: .notice, "Copying build products to \(buildOutputPath.pathString)", metadata: .subsection)
+
+        try FileHandler.shared
+            .contentsOfDirectory(xcodeSchemeBuildPath)
+            .forEach { product in
+                let productOutputPath = buildOutputPath.appending(component: product.basename)
+                if FileHandler.shared.exists(productOutputPath) {
+                    try FileHandler.shared.delete(productOutputPath)
+                }
+
+                try FileHandler.shared.copy(from: product, to: productOutputPath)
+            }
     }
 }